Rename BPM to QPM. Move files from magenta/lib/ to magenta/music/
