+++ magenta/music/musicxml_parser_test.py
@@ -1,5 +1,3 @@
-# Copyright 2016 Google Inc. All Rights Reserved.
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -15,63 +13,51 @@
 
 from collections import defaultdict
 import os.path
+import tempfile
 
 # internal imports
+from musicxml_parser import *
+from musicxml_reader import *
 import tensorflow as tf
-from magenta.common import testing_lib as common_testing_lib
-from magenta.music import sequences_lib
-from magenta.music import testing_lib
-from magenta.music.musicxml_parser import MusicXMLDocument
-from magenta.music.musicxml_reader import musicxml_to_sequence_proto
+
+# self.flute_scale_filename contains an F-major scale of 8 quarter notes each
+
+# self.clarinet_scale_filename contains a F-major scale of 8 quarter notes
+# each appearing as written pitch. This means the key is written as
+# G-major but sounds as F-major. The MIDI pitch numbers must be transposed
+# to be input into Magenta
+
+# self.band_score_filename contains a number of instruments in written
+# pitch. The score has two time signatures (6/8 and 2/4) and two sounding
+# keys (Bb-major and Eb major). The file also contains chords and
+# multiple voices (see Oboe part in measure 57), as well as dynamics,
+# articulations, slurs, ties, hairpins, grace notes, tempo changes,
+# and multiple barline types (double, repeat)
+
+# self.compressed_filename contains the same content as
+# self.flute_scale_filename, but compressed in MXL format
+
 
 class MusicXMLParserTest(tf.test.TestCase):
-  """Class to test the MusicXML parser use cases
-  self.flute_scale_filename contains an F-major scale of 8 quarter notes each
 
-  self.clarinet_scale_filename contains a F-major scale of 8 quarter notes
-  each appearing as written pitch. This means the key is written as
-  G-major but sounds as F-major. The MIDI pitch numbers must be transposed
-  to be input into Magenta
-
-  self.band_score_filename contains a number of instruments in written
-  pitch. The score has two time signatures (6/8 and 2/4) and two sounding
-  keys (Bb-major and Eb major). The file also contains chords and
-  multiple voices (see Oboe part in measure 57), as well as dynamics,
-  articulations, slurs, ties, hairpins, grace notes, tempo changes,
-  and multiple barline types (double, repeat)
-
-  self.compressed_filename contains the same content as
-  self.flute_scale_filename, but compressed in MXL format
-  """
   def setUp(self):
-
-    self.steps_per_quarter = 4
-
     self.flute_scale_filename = os.path.join(
         tf.resource_loader.get_data_files_path(),
-        'testdata/flute_scale.xml')
+        '../testdata/flute_scale.xml')
 
     self.clarinet_scale_filename = os.path.join(
         tf.resource_loader.get_data_files_path(),
-        'testdata/clarinet_scale.xml')
+        '../testdata/clarinet_scale.xml')
 
     self.band_score_filename = os.path.join(
         tf.resource_loader.get_data_files_path(),
-        'testdata/el_capitan.xml')
+        '../testdata/el_capitan.xml')
 
     self.compressed_filename = os.path.join(
         tf.resource_loader.get_data_files_path(),
-        'testdata/flute_scale.mxl')
+        '../testdata/flute_scale.mxl')
 
-    self.multiple_rootfile_compressed_filename = os.path.join(
-        tf.resource_loader.get_data_files_path(),
-        'testdata/flute_scale_with_png.mxl')
-
-    self.rhythm_durations_filename = os.path.join(
-        tf.resource_loader.get_data_files_path(),
-        'testdata/rhythm_durations.xml')
-
-  def checkmusicxmlandsequence(self, musicxml, sequence_proto):
+  def CheckMusicXMLAndSequence(self, musicxml, sequence_proto):
     """Compares MusicXMLDocument object against a sequence proto.
 
     Args:
@@ -79,19 +65,19 @@ class MusicXMLParserTest(tf.test.TestCase):
       sequence_proto: A tensorflow.magenta.Sequence proto.
     """
     # Test time signature changes.
-    self.assertEqual(len(musicxml.get_time_signatures()),
+    self.assertEqual(len(musicxml.getTimeSignatures()),
                      len(sequence_proto.time_signatures))
-    for musicxml_time, sequence_time in zip(musicxml.get_time_signatures(),
-                                            sequence_proto.time_signatures):
+    for musicxml_time, sequence_time in zip(musicxml.getTimeSignatures(),
+                                        sequence_proto.time_signatures):
       self.assertEqual(musicxml_time.numerator, sequence_time.numerator)
       self.assertEqual(musicxml_time.denominator, sequence_time.denominator)
       self.assertAlmostEqual(musicxml_time.time_position, sequence_time.time)
 
     # Test key signature changes.
-    self.assertEqual(len(musicxml.get_key_signatures()),
+    self.assertEqual(len(musicxml.getKeySignatures()),
                      len(sequence_proto.key_signatures))
-    for musicxml_key, sequence_key in zip(musicxml.get_key_signatures(),
-                                          sequence_proto.key_signatures):
+    for musicxml_key, sequence_key in zip(musicxml.getKeySignatures(),
+                                      sequence_proto.key_signatures):
 
       if musicxml_key.mode == "major":
         mode = 0
@@ -107,7 +93,7 @@ class MusicXMLParserTest(tf.test.TestCase):
       self.assertAlmostEqual(musicxml_key.time_position, sequence_key.time)
 
     # Test tempos.
-    musicxml_tempos = musicxml.get_tempos()
+    musicxml_tempos = musicxml.getTempos()
     self.assertEqual(len(musicxml_tempos),
                      len(sequence_proto.tempos))
     for musicxml_tempo, sequence_tempo in zip(
@@ -139,70 +125,28 @@ class MusicXMLParserTest(tf.test.TestCase):
 
       self.assertEqual(len(musicxml_notes), len(seq_instrument_notes))
       for musicxml_note, sequence_note in zip(musicxml_notes,
-                                              seq_instrument_notes):
+                                          seq_instrument_notes):
         self.assertEqual(musicxml_note.pitch[1], sequence_note.pitch)
         self.assertEqual(musicxml_note.velocity, sequence_note.velocity)
-        self.assertAlmostEqual(musicxml_note.note_duration.time_position,
+        self.assertAlmostEqual(musicxml_note.time_position,
                                sequence_note.start_time)
-        self.assertAlmostEqual(musicxml_note.note_duration.time_position \
-                               + musicxml_note.note_duration.seconds,
+        self.assertAlmostEqual(musicxml_note.time_position \
+                               + musicxml_note.seconds,
                                sequence_note.end_time)
-        # Check that the duration specified in the MusicXML and the
-        # duration float match to within +/- 1 (delta = 1)
-        # Delta is used because duration in MusicXML is always an integer
-        # For example, a 3:2 half note might have a durationfloat of 341.333
-        # but would have the 1/3 distributed in the MusicXML as
-        # 341.0, 341.0, 342.0.
-        # Check that (3 * 341.333) = (341 + 341 + 342) is true by checking
-        # that 341.0 and 342.0 are +/- 1 of 341.333
-        self.assertAlmostEqual(musicxml_note.note_duration.duration, \
-                               musicxml_note.state.divisions * 4 \
-                               * musicxml_note.note_duration.duration_float(),\
-                               delta=1)
 
-  def checkmusicxmltosequence(self, filename):
+  def CheckMusicXMLToSequence(self, filename):
     """Test the translation from MusicXML to Sequence proto."""
     source_musicxml = MusicXMLDocument(filename)
     sequence_proto = musicxml_to_sequence_proto(source_musicxml)
-    self.checkmusicxmlandsequence(source_musicxml, sequence_proto)
+    self.CheckMusicXMLAndSequence(source_musicxml, sequence_proto)
 
-  def checkFMajorScale(self, filename):
-    """Verify that the MusicXML scale file contains the correct
-    pitches (sounding pitch) and durations"""
+  def testSimpleMusicXMLToSequence(self):
+    self.CheckMusicXMLToSequence(self.flute_scale_filename)
 
-    # Expected QuantizedSequence
-    # Sequence tuple = (midi_pitch, velocity, start_seconds, end_seconds)
-    expected_quantized_sequence = sequences_lib.QuantizedSequence()
-    expected_quantized_sequence.steps_per_quarter = self.steps_per_quarter
-    expected_quantized_sequence.qpm = 120.0
-    expected_quantized_sequence.time_signature = sequences_lib.QuantizedSequence.TimeSignature(
-      numerator = 4, denominator = 4)
-    testing_lib.add_quantized_track_to_sequence(
-      expected_quantized_sequence, 0,
-      [(65, 64, 0, 4), (67, 64, 4, 8), (69, 64, 8, 12),
-      (70, 64, 12, 16), (72, 64, 16, 20), (74, 64, 20, 24),
-      (76, 64, 24, 28), (77, 64, 28, 32)]
-    )
+  def testComplexMusicXMLToSequence(self):
+    self.CheckMusicXMLToSequence(self.band_score_filename)
 
-    # Convert MusicXML to QuantizedSequence
-    source_musicxml = MusicXMLDocument(filename)
-    sequence_proto = musicxml_to_sequence_proto(source_musicxml)
-    quantized = sequences_lib.QuantizedSequence()
-    quantized.from_note_sequence(sequence_proto, self.steps_per_quarter)
-
-    # Check equality
-    self.assertEqual(expected_quantized_sequence, quantized)
-
-  def testsimplemusicxmltosequence(self):
-    """Test the simple flute scale MusicXML file"""
-    self.checkmusicxmltosequence(self.flute_scale_filename)
-    self.checkFMajorScale(self.flute_scale_filename)
-
-  def testcomplexmusicxmltosequence(self):
-    """Test the complex band score MusicXML file"""
-    self.checkmusicxmltosequence(self.band_score_filename)
-
-  def testtransposedxmltosequence(self):
+  def testTransposedMusicXMLToSequence(self):
     """Test the translation from MusicXML to Sequence proto when the music
     is transposed. Compare a transpoed MusicXML file (clarinet) to an
     identical untransposed sequence (flute)
@@ -210,10 +154,9 @@ class MusicXMLParserTest(tf.test.TestCase):
     untransposed_musicxml = MusicXMLDocument(self.flute_scale_filename)
     transposed_musicxml = MusicXMLDocument(self.clarinet_scale_filename)
     untransposed_proto = musicxml_to_sequence_proto(untransposed_musicxml)
-    self.checkmusicxmlandsequence(transposed_musicxml, untransposed_proto)
-    self.checkFMajorScale(self.clarinet_scale_filename)
+    self.CheckMusicXMLAndSequence(transposed_musicxml, untransposed_proto)
 
-  def testcompressedxmltosequence(self):
+  def testCompressedMusicXMLToSequence(self):
     """Test the translation from MusicXML to Sequence proto when the music
     is compressed in MXL format. Compare a compressed MusicXML file to an
     identical uncompressed sequence
@@ -221,26 +164,7 @@ class MusicXMLParserTest(tf.test.TestCase):
     uncompressed_musicxml = MusicXMLDocument(self.flute_scale_filename)
     compressed_musicxml = MusicXMLDocument(self.compressed_filename)
     uncompressed_proto = musicxml_to_sequence_proto(uncompressed_musicxml)
-    self.checkmusicxmlandsequence(compressed_musicxml, uncompressed_proto)
-    self.checkFMajorScale(self.flute_scale_filename)
-
-  def testmultiplecompressedxmltosequence(self):
-    """Test the translation from MusicXML to Sequence proto when the music
-    is compressed in MXL format with multiple rootfiles. The example
-    MXL file contains a MusicXML file of the Flute F Major scale, as well
-    as the PNG rendering of the score contained within the single MXL file.
-    Compare a compressed MusicXML file to an identical uncompressed sequence
-    """
-    uncompressed_musicxml = MusicXMLDocument(self.flute_scale_filename)
-    compressed_musicxml = MusicXMLDocument(
-                            self.multiple_rootfile_compressed_filename)
-    uncompressed_proto = musicxml_to_sequence_proto(uncompressed_musicxml)
-    self.checkmusicxmlandsequence(compressed_musicxml, uncompressed_proto)
-    self.checkFMajorScale(self.flute_scale_filename)
-
-  def testrhythmdurationsxmltosequence(self):
-    """Test the rhythm durations MusicXML file"""
-    self.checkmusicxmltosequence(self.rhythm_durations_filename)
+    self.CheckMusicXMLAndSequence(compressed_musicxml, uncompressed_proto)
 
 if __name__ == '__main__':
   tf.test.main()
